+++ packages/google-cloud-alloydb/tests/unit/gapic/alloydb_v1alpha/test_alloy_db_admin.py
@@ -25629,6 +25629,7 @@ def test_create_cluster_rest_call_success(request_type):
         },
         "tags": {},
         "service_account_email": "service_account_email_value",
+        "dataplex_config": {"enabled": True},
     }
     # The version of a generated dependency at test runtime may differ from the version used during generation.
     # Delete any fields which are not present in the current runtime dependency
@@ -25919,6 +25920,7 @@ def test_update_cluster_rest_call_success(request_type):
         },
         "tags": {},
         "service_account_email": "service_account_email_value",
+        "dataplex_config": {"enabled": True},
     }
     # The version of a generated dependency at test runtime may differ from the version used during generation.
     # Delete any fields which are not present in the current runtime dependency
@@ -27044,6 +27046,7 @@ def test_create_secondary_cluster_rest_call_success(request_type):
         },
         "tags": {},
         "service_account_email": "service_account_email_value",
+        "dataplex_config": {"enabled": True},
     }
     # The version of a generated dependency at test runtime may differ from the version used during generation.
     # Delete any fields which are not present in the current runtime dependency