+++ packages/google-cloud-documentai/google/cloud/documentai_v1/types/document_processor_service.py
@@ -143,6 +143,12 @@ class ProcessOptions(proto.Message):
             return_bounding_boxes (bool):
                 Optional. Whether to include bounding boxes
                 in layout parser processor response.
+            enable_image_annotation (bool):
+                Optional. Whether to include image
+                annotations in layout parser response.
+            enable_table_annotation (bool):
+                Optional. Whether to include table
+                annotations in layout parser response.
         """
 
         class ChunkingConfig(proto.Message):
@@ -178,6 +184,14 @@ class ProcessOptions(proto.Message):
         return_bounding_boxes: bool = proto.Field(
             proto.BOOL,
             number=3,
+        )
+        enable_image_annotation: bool = proto.Field(
+            proto.BOOL,
+            number=4,
+        )
+        enable_table_annotation: bool = proto.Field(
+            proto.BOOL,
+            number=6,
         )
 
     class IndividualPageSelector(proto.Message):