+++ packages/google-cloud-hypercomputecluster/google/cloud/hypercomputecluster/__init__.py
@@ -18,13 +18,13 @@ from google.cloud.hypercomputecluster import gapic_version as package_version
 __version__ = package_version.__version__
 
 
-from google.cloud.hypercomputecluster_v1beta.services.hypercompute_cluster.async_client import (
+from google.cloud.hypercomputecluster_v1.services.hypercompute_cluster.async_client import (
     HypercomputeClusterAsyncClient,
 )
-from google.cloud.hypercomputecluster_v1beta.services.hypercompute_cluster.client import (
+from google.cloud.hypercomputecluster_v1.services.hypercompute_cluster.client import (
     HypercomputeClusterClient,
 )
-from google.cloud.hypercomputecluster_v1beta.types.hypercompute_cluster import (
+from google.cloud.hypercomputecluster_v1.types.hypercompute_cluster import (
     BootDisk,
     BucketReference,
     Cluster,
@@ -67,8 +67,33 @@ from google.cloud.hypercomputecluster_v1beta.types.hypercompute_cluster import (
     StorageResourceConfig,
     UpdateClusterRequest,
 )
-from google.cloud.hypercomputecluster_v1beta.types.operation_metadata import (
+from google.cloud.hypercomputecluster_v1.types.operation_metadata import (
+    CheckClusterHealth,
+    CreateFilestoreInstance,
+    CreateLoginNode,
+    CreateLustreInstance,
+    CreateNetwork,
+    CreateNodeset,
+    CreateOrchestrator,
+    CreatePartition,
+    CreatePrivateServiceAccess,
+    CreateStorageBucket,
+    DeleteFilestoreInstance,
+    DeleteLoginNode,
+    DeleteLustreInstance,
+    DeleteNetwork,
+    DeleteNodeset,
+    DeleteOrchestrator,
+    DeletePartition,
+    DeletePrivateServiceAccess,
+    DeleteStorageBucket,
     OperationMetadata,
+    OperationProgress,
+    OperationStep,
+    UpdateLoginNode,
+    UpdateNodeset,
+    UpdateOrchestrator,
+    UpdatePartition,
 )
 
 __all__ = (
@@ -115,5 +140,30 @@ __all__ = (
     "StorageResource",
     "StorageResourceConfig",
     "UpdateClusterRequest",
+    "CheckClusterHealth",
+    "CreateFilestoreInstance",
+    "CreateLoginNode",
+    "CreateLustreInstance",
+    "CreateNetwork",
+    "CreateNodeset",
+    "CreateOrchestrator",
+    "CreatePartition",
+    "CreatePrivateServiceAccess",
+    "CreateStorageBucket",
+    "DeleteFilestoreInstance",
+    "DeleteLoginNode",
+    "DeleteLustreInstance",
+    "DeleteNetwork",
+    "DeleteNodeset",
+    "DeleteOrchestrator",
+    "DeletePartition",
+    "DeletePrivateServiceAccess",
+    "DeleteStorageBucket",
     "OperationMetadata",
+    "OperationProgress",
+    "OperationStep",
+    "UpdateLoginNode",
+    "UpdateNodeset",
+    "UpdateOrchestrator",
+    "UpdatePartition",
 )